# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GPU/NPU 服务器管理平台",
    "description": "管理昇腾 NPU 与英伟达 GPU 服务器：状态查看、任务调度、推理结果评测。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/servers": {
      "get": {
        "tags": [
          "servers"
        ],
        "summary": "服务器列表（含卡状态，支持筛选）",
        "operationId": "list_servers_api_servers_get",
        "parameters": [
          {
            "name": "include_devices",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "是否返回每张卡的详情",
              "default": true,
              "title": "Include Devices"
            },
            "description": "是否返回每张卡的详情"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "关键字，匹配名称、地址、主机名、备注",
              "title": "Q"
            },
            "description": "关键字，匹配名称、地址、主机名、备注"
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "运行组，多个为“或”",
              "title": "Group"
            },
            "description": "运行组，多个为“或”"
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "使用人，多个为“或”",
              "title": "Owner"
            },
            "description": "使用人，多个为“或”"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "标签，多个为“且”，即必须同时包含",
              "title": "Tag"
            },
            "description": "标签，多个为“且”，即必须同时包含"
          },
          {
            "name": "accelerator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "gpu",
                    "npu"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "加速卡类型",
              "title": "Accelerator"
            },
            "description": "加速卡类型"
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "加速卡型号，多个为“或”",
              "title": "Model"
            },
            "description": "加速卡型号，多个为“或”"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unknown",
                    "online",
                    "offline"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "在线状态",
              "title": "Status"
            },
            "description": "在线状态"
          },
          {
            "name": "schedulable",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否参与调度",
              "title": "Schedulable"
            },
            "description": "是否参与调度"
          },
          {
            "name": "has_idle",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "true 只返回有空闲卡的服务器",
              "title": "Has Idle"
            },
            "description": "true 只返回有空闲卡的服务器"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerOut"
                  },
                  "title": "Response List Servers Api Servers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "servers"
        ],
        "summary": "添加服务器",
        "operationId": "create_server_api_servers_post",
        "parameters": [
          {
            "name": "deploy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "填写了 ssh_user 时，添加后立即通过 SSH 安装 Agent",
              "default": true,
              "title": "Deploy"
            },
            "description": "填写了 ssh_user 时，添加后立即通过 SSH 安装 Agent"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/grouped": {
      "get": {
        "tags": [
          "servers"
        ],
        "summary": "按属性分组的服务器列表",
        "operationId": "grouped_servers_api_servers_grouped_get",
        "parameters": [
          {
            "name": "by",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "group",
                "owner",
                "tag",
                "accelerator",
                "model",
                "status"
              ],
              "type": "string",
              "description": "分组属性；按标签或型号分组时，一台服务器可出现在多个分组中",
              "title": "By"
            },
            "description": "分组属性；按标签或型号分组时，一台服务器可出现在多个分组中"
          },
          {
            "name": "include_devices",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Devices"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "关键字，匹配名称、地址、主机名、备注",
              "title": "Q"
            },
            "description": "关键字，匹配名称、地址、主机名、备注"
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "运行组，多个为“或”",
              "title": "Group"
            },
            "description": "运行组，多个为“或”"
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "使用人，多个为“或”",
              "title": "Owner"
            },
            "description": "使用人，多个为“或”"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "标签，多个为“且”，即必须同时包含",
              "title": "Tag"
            },
            "description": "标签，多个为“且”，即必须同时包含"
          },
          {
            "name": "accelerator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "gpu",
                    "npu"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "加速卡类型",
              "title": "Accelerator"
            },
            "description": "加速卡类型"
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "加速卡型号，多个为“或”",
              "title": "Model"
            },
            "description": "加速卡型号，多个为“或”"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unknown",
                    "online",
                    "offline"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "在线状态",
              "title": "Status"
            },
            "description": "在线状态"
          },
          {
            "name": "schedulable",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否参与调度",
              "title": "Schedulable"
            },
            "description": "是否参与调度"
          },
          {
            "name": "has_idle",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "true 只返回有空闲卡的服务器",
              "title": "Has Idle"
            },
            "description": "true 只返回有空闲卡的服务器"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerGroup"
                  },
                  "title": "Response Grouped Servers Api Servers Grouped Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/servers/batch": {
      "post": {
        "tags": [
          "servers"
        ],
        "summary": "批量添加服务器",
        "operationId": "batch_create_servers_api_servers_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerBatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerBatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/servers/deploy": {
      "post": {
        "tags": [
          "servers"
        ],
        "summary": "批量安装或升级 Agent",
        "operationId": "deploy_servers_api_servers_deploy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServerOut"
                  },
                  "type": "array",
                  "title": "Response Deploy Servers Api Servers Deploy Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/{server_id}/deploy": {
      "post": {
        "tags": [
          "servers"
        ],
        "summary": "安装或升级该服务器的 Agent",
        "operationId": "deploy_server_api_servers__server_id__deploy_post",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "servers"
        ],
        "summary": "最近一次安装或升级的详情与输出",
        "operationId": "deploy_detail_api_servers__server_id__deploy_get",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DeployDetail"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Deploy Detail Api Servers  Server Id  Deploy Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/servers/{server_id}/refresh": {
      "post": {
        "tags": [
          "servers"
        ],
        "summary": "立即拉取一次该服务器状态",
        "operationId": "refresh_server_api_servers__server_id__refresh_post",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServerBatchCreate": {
        "properties": {
          "servers": {
            "items": {
              "$ref": "#/components/schemas/ServerCreate"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Servers"
          },
          "deploy": {
            "type": "boolean",
            "title": "Deploy",
            "description": "添加后立即通过 SSH 安装 Agent（只对填写了 ssh_user 的服务器）",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "servers"
        ],
        "title": "ServerBatchCreate"
      },
...
      "ServerGroup": {
        "properties": {
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key",
            "description": "分组值，为 null 表示未设置该属性"
          },
          "server_count": {
            "type": "integer",
            "title": "Server Count"
          },
          "online_count": {
            "type": "integer",
            "title": "Online Count"
          },
          "device_count": {
            "type": "integer",
            "title": "Device Count"
          },
          "idle_device_count": {
            "type": "integer",
            "title": "Idle Device Count"
          },
          "servers": {
            "items": {
              "$ref": "#/components/schemas/ServerOut"
            },
            "type": "array",
            "title": "Servers"
          }
        },
        "type": "object",
        "required": [
          "key",
          "server_count",
          "online_count",
          "device_count",
          "idle_device_count",
          "servers"
        ],
        "title": "ServerGroup"
      },
...
    }
  }
}